requirements_dev.txt
# Dependencies used for setting up a venv for development.
# These should be pinned to specific versions.

black==22.6.0
flake8==5.0.4

pylint==2.14.5
astroid==2.11.7

scriv==0.16.0
bump2version==1.0.1

pytest==7.1.3
pytest-cov==4.1.0
pytest-xdist==3.3.1
hypothesis[numpy,pandas]==6.54.6
coverage==7.0.5

py-evm==0.7.0a4 # this version compatible with boa 0.1.7
vyper==0.3.9
titanoboa==0.1.7  # 0.1.6 for python versions before 3.10

sphinx==5.3.0 
Pallets-Sphinx-Themes==2.0.3

mypy==1.5.1
pandas-stubs==2.1.4.231227
